+++ gulpfile.js
@@ -15,7 +15,9 @@ const config = {
     './assets/js/custom_config_mercadopago.js',
     './assets/js/ticket_config_mercadopago.js',
     './assets/js/ticket.js',
+		'./assets/js/pix_config_mercadopago.js',
     './assets/js/review.js',
+    './assets/js/pix_mercadopago_order_received.js',
   ],
   stylesheets: [
     './assets/css/admin_notice_mercadopago.css',