+++ includes/notification/class-wc-woomercadopago-notification-ipn.php
@@ -127,7 +127,10 @@ class WC_WooMercadoPago_Notification_IPN extends WC_WooMercadoPago_Notification_
 				$order->update_meta_data( __( 'Buyer email', 'woocommerce-mercadopago' ), $data['payer']['email'] );
 			}
 			if ( ! empty( $data['payment_type_id'] ) ) {
-				$order->update_meta_data( __( 'Payment method', 'woocommerce-mercadopago' ), $data['payment_type_id'] );
+				$order->update_meta_data( __( 'Payment type', 'woocommerce-mercadopago' ), $data['payment_type_id'] );
+			}
+			if ( ! empty( $data['payment_method_id'] ) ) {
+				$order->update_meta_data( __( 'Payment method', 'woocommerce-mercadopago' ), $data['payment_method_id'] );
 			}
 			if ( ! empty( $data['payments'] ) ) {
 				$payment_ids = array();
@@ -153,7 +156,10 @@ class WC_WooMercadoPago_Notification_IPN extends WC_WooMercadoPago_Notification_
 				update_post_meta( $order->id, __( 'Buyer email', 'woocommerce-mercadopago' ), $data['payer']['email'] );
 			}
 			if ( ! empty( $data['payment_type_id'] ) ) {
-				update_post_meta( $order->id, __( 'Payment method', 'woocommerce-mercadopago' ), $data['payment_type_id'] );
+				update_post_meta( $order->id, __( 'Payment type', 'woocommerce-mercadopago' ), $data['payment_type_id'] );
+			}
+			if ( ! empty( $data['payment_method_id'] ) ) {
+				update_post_meta( $order->id, __( 'Payment method', 'woocommerce-mercadopago' ), $data['payment_method_id'] );
 			}
 			if ( ! empty( $data['payments'] ) ) {
 				$payment_ids = array();