+++ includes/payments/class-wc-woomercadopago-custom-gateway.php
@@ -38,7 +38,7 @@ class WC_WooMercadoPago_Custom_Gateway extends WC_WooMercadoPago_Payment_Abstrac
 		$this->form_fields        = array();
 		$this->method_title       = __( 'Mercado Pago - Custom Checkout', 'woocommerce-mercadopago' );
 		$this->title              = __( 'Pay with debit and credit cards', 'woocommerce-mercadopago' );
-		$this->method_description = $this->get_method_mp_description( $this->description );
+		$this->method_description = $this->description;
 		$this->coupon_mode        = $this->get_option_mp( 'coupon_mode', 'no' );
 		$this->field_forms_order  = $this->get_fields_sequence();
 		parent::__construct();