+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [5.1.0] - 2021-03-29
+
+### Added
+ - Added new Pix Gateway for Brazil
+ - Added Payment type at order panel
+### Changed
+ - Fixed post in configuration page, removed html
+ 
 ## [5.0.1] - 2021-03-10
 ### Added
  - Compatibility with old notification urls